[IIIF-913] Properly check for duration in AbstractCanvas.toJSON (#45)

+++ src/main/java/info/freelibrary/iiif/presentation/v3/AbstractCanvas.java
@@ -407,7 +407,7 @@ abstract class AbstractCanvas<T extends AbstractCanvas<T>> extends NavigableReso
                 json.remove(Constants.HEIGHT);
             }
         }
-        if (json.containsKey(Constants.DURATION) && json.getInteger(Constants.DURATION) == 0) {
+        if (json.containsKey(Constants.DURATION) && json.getFloat(Constants.DURATION) == 0.0f) {
             json.remove(Constants.DURATION);
         }
